+++ test/specs/openapi/test_examples.py
@@ -430,7 +430,7 @@ def test_partial_examples(empty_open_api_3_schema):
     strategy = operation.get_strategies_from_examples()[0]
     # Then all generated examples should have those missing parts generated according to the API schema
     example = get_single_example(strategy)
-    parameters_schema = parameters_to_json_schema(operation.path_parameters)
+    parameters_schema = parameters_to_json_schema(operation, operation.path_parameters)
     jsonschema.validate(example.path_parameters, parameters_schema)